+++ src/modules/mavlink/mavlink_messages.cpp
@@ -73,13 +73,11 @@
 #include <uORB/topics/input_rc.h>
 #include <uORB/topics/manual_control_setpoint.h>
 #include <uORB/topics/manual_control_switches.h>
-#include <uORB/topics/position_controller_status.h>
 #include <uORB/topics/position_setpoint_triplet.h>
 #include <uORB/topics/sensor_baro.h>
 #include <uORB/topics/sensor_gps.h>
 #include <uORB/topics/sensor_mag.h>
 #include <uORB/topics/sensor_selection.h>
-#include <uORB/topics/tecs_status.h>
 #include <uORB/topics/telemetry_status.h>
 #include <uORB/topics/transponder_report.h>
 #include <uORB/topics/vehicle_air_data.h>
@@ -121,6 +119,7 @@ using matrix::wrap_2pi;
 #include "streams/HIGH_LATENCY2.hpp"
 #include "streams/HIL_STATE_QUATERNION.hpp"
 #include "streams/MOUNT_ORIENTATION.hpp"
+#include "streams/NAV_CONTROLLER_OUTPUT.hpp"
 #include "streams/OBSTACLE_DISTANCE.hpp"
 #include "streams/OPTICAL_FLOW_RAD.hpp"
 #include "streams/ORBIT_EXECUTION_STATUS.hpp"
@@ -3776,82 +3775,6 @@ protected:
 	}
 };
 
-class MavlinkStreamNavControllerOutput : public MavlinkStream
-{
-public:
-	const char *get_name() const override
-	{
-		return MavlinkStreamNavControllerOutput::get_name_static();
-	}
-
-	static constexpr const char *get_name_static()
-	{
-		return "NAV_CONTROLLER_OUTPUT";
-	}
-
-	static constexpr uint16_t get_id_static()
-	{
-		return MAVLINK_MSG_ID_NAV_CONTROLLER_OUTPUT;
-	}
-
-	uint16_t get_id() override
-	{
-		return get_id_static();
-	}
-
-	static MavlinkStream *new_instance(Mavlink *mavlink)
-	{
-		return new MavlinkStreamNavControllerOutput(mavlink);
-	}
-
-	unsigned get_size() override
-	{
-		return (_pos_ctrl_status_sub.advertised()) ?
-		       MAVLINK_MSG_ID_NAV_CONTROLLER_OUTPUT_LEN + MAVLINK_NUM_NON_PAYLOAD_BYTES : 0;
-	}
-
-private:
-	uORB::Subscription _pos_ctrl_status_sub{ORB_ID(position_controller_status)};
-	uORB::Subscription _tecs_status_sub{ORB_ID(tecs_status)};
-
-	/* do not allow top copying this class */
-	MavlinkStreamNavControllerOutput(MavlinkStreamNavControllerOutput &) = delete;
-	MavlinkStreamNavControllerOutput &operator = (const MavlinkStreamNavControllerOutput &) = delete;
-
-protected:
-	explicit MavlinkStreamNavControllerOutput(Mavlink *mavlink) : MavlinkStream(mavlink)
-	{}
-
-	bool send() override
-	{
-		if (_pos_ctrl_status_sub.updated()) {
-
-			position_controller_status_s pos_ctrl_status{};
-			_pos_ctrl_status_sub.copy(&pos_ctrl_status);
-
-			tecs_status_s tecs_status{};
-			_tecs_status_sub.copy(&tecs_status);
-
-			mavlink_nav_controller_output_t msg{};
-
-			msg.nav_roll = math::degrees(pos_ctrl_status.nav_roll);
-			msg.nav_pitch = math::degrees(pos_ctrl_status.nav_pitch);
-			msg.nav_bearing = (int16_t)math::degrees(pos_ctrl_status.nav_bearing);
-			msg.target_bearing = (int16_t)math::degrees(pos_ctrl_status.target_bearing);
-			msg.wp_dist = (uint16_t)pos_ctrl_status.wp_dist;
-			msg.xtrack_error = pos_ctrl_status.xtrack_error;
-			msg.alt_error = tecs_status.altitude_filtered - tecs_status.altitude_sp;
-			msg.aspd_error = tecs_status.airspeed_filtered - tecs_status.airspeed_sp;
-
-			mavlink_msg_nav_controller_output_send_struct(_mavlink->get_channel(), &msg);
-
-			return true;
-		}
-
-		return false;
-	}
-};
-
 class MavlinkStreamCameraCapture : public MavlinkStream
 {
 public:
@@ -3988,7 +3911,9 @@ static const StreamListItem streams_list[] = {
 #if defined(DEBUG_FLOAT_ARRAY_HPP)
 	create_stream_list_item<MavlinkStreamDebugFloatArray>(),
 #endif // DEBUG_FLOAT_ARRAY_HPP
+#if defined(NAV_CONTROLLER_OUTPUT_HPP)
 	create_stream_list_item<MavlinkStreamNavControllerOutput>(),
+#endif // NAV_CONTROLLER_OUTPUT_HPP
 	create_stream_list_item<MavlinkStreamCameraCapture>(),
 	create_stream_list_item<MavlinkStreamCameraTrigger>(),
 	create_stream_list_item<MavlinkStreamCameraImageCaptured>(),
